+++ nautilus_trader/common/config.py
@@ -35,6 +35,7 @@ from nautilus_trader.model.data import BarType
 from nautilus_trader.model.enums import AccountType
 from nautilus_trader.model.enums import BookType
 from nautilus_trader.model.enums import OmsType
+from nautilus_trader.model.enums import OtoTriggerMode
 from nautilus_trader.model.enums import TriggerType
 from nautilus_trader.model.identifiers import ComponentId
 from nautilus_trader.model.identifiers import Identifier
@@ -121,7 +122,7 @@ def msgspec_encoding_hook(obj: Any) -> Any:  # noqa: C901 (too complex)
         return str(obj)
     if isinstance(obj, (Price | Quantity | Money | Currency)):
         return str(obj)
-    if isinstance(obj, (OmsType | AccountType | BookType)):
+    if isinstance(obj, (OmsType | AccountType | BookType | OtoTriggerMode)):
         return obj.name
     if isinstance(obj, (pd.Timestamp | pd.Timedelta)):
         return obj.isoformat()
@@ -161,6 +162,8 @@ def msgspec_decoding_hook(obj_type: type, obj: Any) -> Any:  # noqa: C901 (too c
         return AccountType[obj]
     if obj_type == BookType:
         return BookType[obj]
+    if obj_type == OtoTriggerMode:
+        return OtoTriggerMode[obj]
     if obj_type == TriggerType:
         return TriggerType[obj]
     if obj_type == Environment: